Drop TaskExample generic.
